- moves extensions methods in their own namespace to address this comment https://github.com/microsoft/kiota/pull/38#discussion_r593426255
