+++ src/Kiota.Builder/Extensions/OpenApiReferenceExtensions.cs
@@ -1,6 +1,6 @@
 using Microsoft.OpenApi.Models;
 
-namespace Kiota.Builder {
+namespace Kiota.Builder.Extensions {
     public static class OpenApiReferenceExtensions {
         public static string GetClassName(this OpenApiReference reference) {
             var referenceId = reference?.Id;
