+++ src/Kiota.Builder/Extensions/StringExtensions.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Linq;
 
-namespace Kiota.Builder {
+namespace Kiota.Builder.Extensions {
     public static class StringExtensions {
         public static string ToFirstCharacterLowerCase(this string input)
             => string.IsNullOrEmpty(input) ? input : $"{char.ToLowerInvariant(input.FirstOrDefault())}{input.Substring(1)}";
