+++ src/Kiota.Builder/KiotaBuilder.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Readers;
+using Kiota.Builder.Extensions;
 
 namespace Kiota.Builder
 {
