+++ src/Kiota.Builder/OpenApiUriSpaceNode.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using Microsoft.OpenApi.Models;
+using Kiota.Builder.Extensions;
 
 namespace Kiota.Builder
 {
