+++ src/Kiota.Builder/Writers/TypeScriptWriter.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Kiota.Builder.Extensions;
 
 namespace Kiota.Builder
 {
